moved resultsRenderer.ts
